Make Depths exit a regular branch stair.
This might be a bit debatable, but to me having a portal both ways suggests
you're magically transported somewhere.  A mundane stair protected by a gate
would have a gate on one side and a regular stairway on the other -- a pair
of gates is possible but would be unnecessary expense on part of the
dungeon's builders.