Tinker with Depths monster spawn weights.
This is intended to shuffle some of the exp to later in Depths; some of
the mid-tier monsters for the area (mostly giants) are weighted to
appear more in mid-Depths than have a flat spawn rate, and some of the
lower-tier monsters are weighted to appear more and earlier. Overall
there's a slight reduction in the amount of exp available overall, but
it's only a 1-2% change on average.

There are a couple of additions here (suggested by |amethyst) that I'm
not exactly convinced about yet, but we'll see how these fly (heh).