Don't place 1s in the Depths through Pan/Hell entries.
This fixes the chance of Pan entries spawning anything from imps to
Fiends; instead it picks a 2 to place over the gate (tougher than the
rabble that surrounds it, but hopefully not insurmountably difficult).

This also forces a monster set for a Hell entry so that it doesn't spawn
randomly excessive creatures.