Simplify some Snake monster distributions.
A FLAT rectangle plus PEAK of the same amplitude is exactly same as SEMI
with amplitude of the sum of those two.  FLAT + UP/DOWN with a peak at
an end of the range is same as SEMI with some irrelevant junk outside the
range.

What's the point of guardian serpents having a miniscule FLAT in addition
to a PEAK almost ten times larger?  It's not like people really see minor
details like this.