Asterion, a late-game unique minotaur of Makhleb
He is themed as a bloodthirsty ruler cursed to become a minotaur who now
wanders the dungeon in the service of Makhleb and hoping to establish a
new kingdom in its depths.

The name is taken from ancient greek mythology, which was either the
name of the minotaur itself or of a king related to the minotaur:

http://en.wikipedia.org/wiki/Asterion

Thanks to SGrunt (the S is silent) for the idea of the name.

Stat-wise, Asterion is a minotaur, but with 15 HD, 120 HP, a 1d40
primary attack (instead of 1d35), and the only resistances I've added
are Sinv and high MR.  Experience and placement-wise he's the same as
Frances. He always gets a good demon weapon, large shield, and a robe.

Spell-wise, he has Makhleb's major destruct and spectral weapon.  The
major destruct code is basically identical to the player version: it
chooses randomly from 7 beam types, including orb of electricity, with a
range of 7.

[committer's note: Changed to ETC_FIRE (Makhleb altar colour) to
distinguish from plain minotaurs. There is a proposal to move tengu
thereby freeing up space on 'H', but I didn't want to tie this commit
to that change. Also improved the quote formatting and citation. -nfm]