Don't claim that Vaults spawn on D:2-6.
In transferred saves, we don't have information about the depth range used,
so the range comes from the current version of the game -- ie, it describes
what would happen if you started a new game right now.  On the other hand,
the parent branch came from the old version.  Be consistent and knowingly
provide false data -- this is less jarring than being truthful for the
branch but wrong about depths.