Equip minotaurs that naturally appear outside of labyrinths.
They appear as mercenaries, in a handful of vaults and are currently
spawning in Depths; in those circumstances, you'd expect them to bring
their own combat equipment.

Right now they're equipped similarly to orc warlords, vault wardens, and
other high-tier equipment-using monsters, though this borrows the idea
originating in a2e80fb to occasionally give them a rod to use (as though
they picked it up in a labyrinth). It is really low weight right now and
might go away entirely if it doesn't prove to be very interesting and/or
generates too many rods to balance their power.