Fix an unlinked items bug with corruption and door sealing.
move_items in items.cc was mishandling the special case in which the
source square is empty. In this case, it would delink all the items in the
destination square.

There are several parts of the code that call move_items, but only the
corruption code and the vault warden door sealing code call it without
first ensuring there are items to move.

To trigger this bug: 1. Use corruption in such a way that a monster standing
on an empty square gets displaced by a forming wall onto a square with an
item in it. 2. Stand on a door, make sure there are no items under you,
have the warden seal the door, displacing you onto an item.