Implement adjacent_iterator on its own, rather than via radius_iterator.
This hardly speeds up things, which I did not know as profile runs did not
help to tell these two apart and it turns out adjacent_iterator is not that
prominent...  but hey, a 0.1% speedup for two pages of code might still be
better committed than not.

This speedup might disappear when I get around to the actual rewrite of
radius_iterator, though.