Have ENCH_BREATH_WEAPON interrupt / prevent Word of Recall.
Also add ENCH_BREATH_WEAPON after using Word of Recall.

It's a long recitation; similar to the player's use of Recite it makes
sense for it to wind the caster.

The two immediate gameplay impacts are that 1) it can't be spammed (if
a monster capable of doing so ever comes into existence) and 2) curare
interrupts Word of Recall.