Force a "no" answer when yesno() gets cancelled by a HUP.
This is wrong: if there's indeed no correct default answer, the caller
should reissue the question when the game is loaded.  Not every caller
does, but it seems all cases yet unhandled are:
* unidentified scroll of brand weapon
* unidentified scroll of ID/EA/charging
* any scroll of ID on 2nd or 3rd item

It's likely the player will start thinking, read learndb/spoilers, etc,
when he gets branding, which can result in a HUP -- but that's still
better than corrupting saves and/or allowing cheats.