Remove unused Xom greater demon summoning code
Per #7321, Xom can't ever actually summon holies or greater demons, and
this appears to have been the case for years (9e6fb5de4cf9 is the culprit
as far as I can tell). It might be good to expand Xom's list of summons
from just demons, but it's probably better to just do so from scratch
rather than attempt to re-use ancient code for it.