Don't compute slime neighbours on levels with no slime.
2% of a qw profile was spent there.

Might be better to optimize it for levels with actual slime on them as well:
we call set_terrain_changed() which can update slime neighbour counts,
removing the need to compute them repeatedly.  Not doing this right now only
because I'm not sure about all cases vault placement can alter the terrain.