Axe a bunch of redundant arena sprint renames.
Especially boring ones that come together with an interesting one, such as
the succubus or the witch (a novel name, a set of spells, a desc and a
tile).  It's rare anyone gets that far, so they just watered down better
renames that already compete with regular uniques.

Even with this culling, your chances to meet a succubus are 1/6 for a game
that's already most of its way towards winning.