Add hit_(non)magic(_nomove) functions, rename hit_adjacent to hit_closest_nomove.
One issue with the recently added automagic functionality is that it
displaces the regular tab autofight, making it painful to use for characters
who would still like to melee weak monsters.

For now, I'm leaving the default automagic behavior as it is for a bit more
testing and feedback. This commit makes it easier for players to put autofight
and automagic on two different keys, by macroing one to ===hit_nonmagic and
the other to ===hit_magic. (I plan to use tab and capslock, personally.)
There are also non-moving versions at hit_nonmagic_nomove() and
hit_magic_nomove().

For more consistent and correct naming, hit_adjacent() (shift-tab by default)
was renamed to hit_closest_nomove().