Cap monster summons.
This has most noticeably been an issue with, say, Ilsuiw's water
elementals (or old-and-then-new-again elemental wellsprings) and
Sojobo's air elementals; it tended to get extremely spammy and tedious.

There are some monster spells for which I've set the cap fairly high -
perhaps unnecessarily so; these can be tinkered with as necessary.