Rework the "extra" tag for non-minivaults.
Instead of mixing extra vaults in with primary vaults, a primary vault
without the extra tag is placed first, and vaults with ORIENT: tags and
the extra tag are placed last in a manner similar to how abyssal vaults
are handled

Handling of minivaults is unaffected.

This is intended to prevent Depths from placing two huge scary primary
vaults on one level, but it also has handling relevant for upcoming
tweaks to areas of the dungeon currently dependent on tagged vaults.