Remove the "pan" vault tag; improve Pan builder handling of vaults.
The primary vault on a Pan level is now always a pan lord vault, whether
that be a major or minor pan lord. Pan lord vaults thus get ORIENT: tags
(and no extra tag).

This depends on the handling of the previous commit - previously it was
possible for picking a map by depth to return a map tagged "extra"; we
don't want that here.