Remove M_FLEES from dragons and drakes
This behavior makes this monsters annoying to fight.
I believe the intention behind this flag is to cause them to step
back and blast you with their breath weapon. In practice this
doesn't happen. Instead the player is forced to run them down. If
you aren't speed 10, this is really no fun.

If we want to put this behavior back in, I strongly recommend
that we modify it so dragons would quit fleeing as soon as they
can get a good blast off at the player.