Remove Dragonform Knockback
Conceptually knockback is a cool idea. It's consistent with
the monster behavior, it's thematic, it makes you feel like
a tough guy (for about three hits). After that, it starts
to become a massive inconvenience. We could add a kludge to
allow toggling it, but I don't see how this would improve
gameplay.

Monster dragon knockback is mostly useful for blundering
players off of stairs. Since monsters don't stairdance you,
this is mostly useless as a player ability.

Here are a couple of ideas if we want to retain the flavor
of dragons being tough and awesome:
* Knockback _without_ following (perhaps with stun?)
* Stunning
* A chance to do massive damage