Drop LTO on arches other than *amd64*, *i386*, *x32*.
(kfreebsd-amd64 or even hurd-i386 are ok.)

Crawl's code base is so hard on compilers it probably should be put as-is
into gcc's and clang's test suites.  There's _always_ a build failure on
at least one arch somewhere, and I can't do serious bug reporting on
architectures I can't build on reasonably fast.

On arm*, this fails due to:
https://sourceware.org/bugzilla/show_bug.cgi?id=15639