Revert "Delay more prompt until level has finished loading."
While the problem this commit attempts to fix is real, I'm afraid it causes
more breakage:
* it removes debug messages from map generation
  * especially, infinite loops become extremely hard to debug
* loses output that overflows
* has issues with force_more
* portal vault announcements are lost

For example, a level with both a Lab and an Ice Cave shows just:
"There is an entrance to an ice cave on this level. Hurry and find it before the portal melts!"
even though there's no overflow.  The rest (the main "wave of frost" message
and anything about the Lab) are gone.

I'm afraid some other approach is needed.  Perhaps forcing a screen draw?

This reverts commit 8471a022e76bae39918ba6eaa1fbbab5c7ceb64f.