Ensure an accessible path through hall_of_Zot.
This is intended to address reports of having two Zot traps blocking the
two lungs off; I'm not convinced that this is a good solution, or how
much of a problem this is to begin with, but it's here for now.

Note: this doesn't prevent mechanical traps from placing across both
entrances, but having it completely blocked off by two teleport traps,
Zot traps, etc. should no longer be possible.