Unsplit an URL.
I do need to scroll to copy it in chunks (or cat it first), but that's
hardly any more work when it's in one piece as compared to being chopped.
Those with a GUI editor, on the other hand, can now click it directly.

Also, write those accents in random directions inline -- I don't think
anyone these days can be expected to have problems with non-ASCII.