More tutorial text fixes in translations.
I fixed the formatting as best as I could, but not understanding the
languages, I surely did some mistakes. I especially didn't try to fix the
"tutorial1 levelmap" entry as I would probably just make it worse.

For the languages with weird whitespace rules (ja, ko and zh), I only tried
to make sure that there was no broken tag, but I didn't fix the whitespaces.
The result is a bit ugly and inconsistent, but (hopefully) not buggy.
I leave it to translators to finish the formatting properly.

The wrapper still has some quirks with those languages (adding whitespace,
putting a dot alone on a line,...) but at least it seems it doesn't break
tags anymore.