Lower summoning caps by 1 to where they should be.
They were, apparently inadvertently, increased in 0.14-a0-1164-g531d946,
when the summon cap loop was made to skip the new summon (so that "3"
meant "3 plus the new one").  Add one to the count to compensate.

Also, revert the part of 0.14-a0-1339-gd040c24 that worked around this
problem for CCF.