Allow non-adjacent non-religious followers to take stairs.
This allows abominations from Twisted Resurrection to follow you en
masse, without the tedium of tf>tw...<tf>tw... etc.  It also affects
mercenaries and other stair-taking allies.

TR is, despite changes in 0.10, still one of those spells that is
overpowered but rarely-used because of tedium.  Let's reduce the tedium,
then nerf in other ways as necessary.  Non-healing will be one part of
the nerf, but we probably also need a cap on the total number of
abominations.