[txc] Allow breaking lines between 2 tags.
This helps readability quite a lot. Although it means that putting a whitespace
between 2 tags should be avoided when possible, because it might be removed
on unwrap (both by txc and crawl), if the line break happens to be on this
specific whitespace. Quite unlikely, but if it does happen, it can be fixed by
using a non-breaking space (0xA0).

Note that it works by inserting a line feed (\f) between tags which is later
removed. This makes lines containing consecutive tags to have a slightly
shorter maximal length.