Don't get stuck in autoexplore (#7855)
If the player is standing on an autopickup item and greedy explore is
on, the following situation (where ? is unexplored) reliably causes
autopickup to report "Can't reach some places".

 ####
 ##.#
 #@#?
 .###

We decide in the pathfinding code that NE (being adjacent to an
unexplored square) is a reasonable exploration destination, so return
true from path_examine_point for that square.  This happens before we
explore back to our starting square, even though greedy_place is already
set to that location (by point_traverse_delay?!).  Since the greedy
square is closer, explore_target() and thus pathfind() returns that
square (the player's location).  But then _explore_find_target_square
rejects the square because it wasn't actually visited by pathfind and
thus appears to be unreachable.

Avoid this situation by special-casing the player's location in
_explore_find_target_square, so that it isn't considered unreachable.
That might not be the ultimately best approach, but it seems the safest
for now because it doesn't risk subtly breaking other uses of pathfind.