Refactor colouring of features.
This allows defining just the colour without having to repeat all other
properties.  Apply such colours to portal vault entrances and exits -- using
whatever the old code tried to get.  This commit merely restores that intent,
it would be nice to colour missing ones in a way that makes sense.

Only entrances and exits are deduplicated, it would be nice to apply this
to other features too.