Drop 'l' monsters from Snake again.
It appears that they're merely an annoyance, while not bringing anything
new.  Neither komodo dragons, lindwurms nor basilisks are snakes of any
kind, so they don't fit in theme-wise as well.

During the discussion on ##crawl, the reason they were added was quoted
as "because one vault has lindwurms", to which people responded that the
vault in question (snake_pit_salamanders_mu) feels problematic by itself.