Improve Corpse Rot messaging.
It previously gave the "you smell decay" line if you could smell, and
could see the caster.  That means that if you were invisible and cast
the spell, there would be no message.

Now give the message whenever you are the caster, can see the caster, or
can see a resulting miasma cloud.  If you can't smell, give a message
that doesn't involve smell.