Ignore brentry when returning from a ziggurat.
The code checked this before the level stack, thus if you returned from Pan,
a trovel or (in transferred saves) D, you got placed in Depths instead.
For extra fun, try this in a pre-Depths save! :þ

We could just check the stack first, but the current code allows for a
regular subbranch of the Abyss, etc.