Prompt to continue wearing/unwearing armour when a monster comes into view.
Without this, you can easily die if a hydra or other nasty thing comes
into view after you start equipping or removing heavy body armour.
Regardless of whether this is an rcfile option, it is apparently
not very well-known, and it is a better default than the one in place.
I tried to fix memorization and multidrop in a similar fashion, but
adding monster to the relevant places prevented you from attempting
either action, even if a only a goblin came into view.