Default 'note_chat_messages' to false.
I believe that most players actually want this option disabled by default.
With it on, chat messages get written out into morgues. This leads to a lot
of banal chatter getting dumped. The last time I killed this feature off it
was because at least one person was being creepy with the chat logs. While
players can log whatever shows up if they so desire, I think there's a
significant difference between this behavior and the server logs attesting
to the authenticity of a message by writing it into the morgue.