Default language to the system locale [Unix].
This should increase the amount of players who actually get to see
translations by 2-3 orders of magnitude.  I did not do this initially
on purpose, not wanting to make push jarring work in progress.  The
progress of translating goes way, way slower than anticipated, though,
so enabling just descriptions is not that bad.

There are three platforms that need to be handled
* Unix (this commit)
* Windows
* webtiles