Display minimum delay in a weapon's description.
It could be easily calculated, and it was usually considered extremely
important while only being mentioned in passing in the manual. Thus, all
three important delays are visible now: mindelay, current delay (by
swinging your weapon), and base delay. Like other things on the screen,
it's a base property of the weapon, so there is no reason not to show
it.

The screen had to be condensed a little to fit in 80 chars, but even the
dark maul made it.