Identify last unknown targetted scroll and healing potion
... when you know what they are.

Specifically, you know what they are when it has been tried (so you know it
is in one of those categories) and when all the others in that category are
identified. This should probably be extended to nontargetted scrolls, mainly
for ?CW for Mo, and for the last unidentified scroll and last unidentified
potion.

[With spelling changes to reflect 0.14-a0-1407-g9ff65a6 -nfm]