Remove `pickup_mode` option, introduce `pickup_menu`
End result is that:
* `pickup_menu = true` is old `pickup_mode = multi`, and new default
* `pickup_menu = false` is old `pickup_mode = single`
* `pickup_mode = auto:X` is gone without replacement for being of little use

Recognizing the old option in initfile did not make much sense either
because the default changed: Old `multi` was the only setting that
anybody would include in their rcfile, and that one is now meaningless.

So if your config until now included a line like the following:
    pickup_mode = multi
you can now safely remove this.
If you prefer the old default behavior, use this option from now on:
    pickup_menu = false