Use -isystem instead of -I for (what should be) system libraries
This includes contribs.

We don't do this for system sqlite, though, since that's expected to
be either in /usr/include et al or from contribs, and explicitly
passing "-isystem /usr/include" screws up the include path ordering.

(Instead, we drop the -I flag for sqlite entirely.  If it turns out to
be needed somewhere, the comments were wrong ...)