Allow delayed_decay to work with , and / itemspecs.
Several existing calls used something like:
  dgn.delayed_decay(_G, '%', 'human corpse / human skeleton')
But that actually only added never_decay to the corpse, not the
skeleton.  Now add delayed_

Also add a new function dgn.delayed_decay_extra, like delayed_decay
but taking one additional argument: a string listing additional
non-corpse items to place on the spot.  If the argument is not nil,
it is appended to the itemdef with a comma.  Use this for cases
like 'human corpse, battleaxe'.

Finally, for the few cases that used something like "human corpse / nothing",
use a SUBST to get the same weights.