Move some non-demons (unseen horror, uglies) away from mon/demons/
It's a common mistake for players, too -- nothing tells you you can vamp
drain or agony these, and learning that is quite a surprise.  I guess we'd
better use that proposed "eldritch" holiness for them, making them
undrainable.

The directory name sucks and is a D&Dism -- sadly, "eldritch" is still not
right as without the word having an in-game meaning it's mostly a synonym
for "abyssal" which these critters are not.