Make salamanders amphibious wrt lava / spawn occasionally in late Snake.
This has been suggested a few times. Gameplay-wise, they're not very
challenging to avoid (except for the Snake ending that places them)
thanks to the constraint of being stuck in a pool of lava; flavour-wise,
the mythological creature it's based on is amphibious in many
representations.

Snake could use a slightly more diverse enemy set anyway (see
snake-enemies branch).

This introduces an HT_AMPHIBIOUS_LAVA, which is exactly what it sounds
like; perhaps a better name could be found for it.