Remove a Xom trap vault
It's incredibly spoilery: if you go near it with a monster in LOS, the
item is destroyed, presumably to prevent monsters being used to pick up
the wand and move it away from the trigger? Even if the item were marked
no_pickup to solve this, handing out items like wands of teleportation in
a spoilery way isn't great.