Shadow god: Shadow Step.
Within a piety-dependent range (3 at 2* up to LOS at 6*), you can move
into the shadow of a nearby creature assuming it has a shadow (i.e. not
invisible and one other condition) and that it's sufficiently still for
you to target it - i.e. the target is immobile in one form or another.

This is basically a jump attack without the attack and additional
conditions, and apes a lot of the code thereof.

Cost is currently 4 MP and 4 piety (piety cost is the same as Finesse,
Imprison, and Rain amongst other abilities).