Shadow god: Bleed Smoke.
Always triggers on a 50%-mhp ("Ouch!") attack; the lower bound goes down
linearly to 0 damage at 200 piety, with a linearly-scaling maximum
chance over that range (so at 200 piety a 25%-mhp attack has a 50%
chance to trigger fog).

The cloud is half the size as you get with a scroll of fog or the Cloak
of the Thief (though it seems to spread fairly rapidly).

The specific numbers here might need to be adjusted more to balance it
out properly.