Remove caps; rename wizards hats "hats."
In the spirit of the glove/gauntlet and helmet types distinction removal, there
doesn't seem to be much reason to have both caps and wizard hats. Hats are more
common among unrands and wizards starting with hats is more "thematic" than
some gladiators starting with caps. on the other hand, those gladiators starting
with wizard hats would be pretty weird, so wizard hats are just hats now.

Caps not removed:
- Caps existing in old saves
- Summon cap
- Level cap

This changes the percentages of hat brands from 2/3 MR and 1/3 int to 1/3 MR,
1/3 int, and 1/3 spirit shield.