Adjustments to Shadow Form.
Miasma is gone; miasma is more strongly associated with necromancy
than anything related to darkness or shadows - see e.g. Corpse Rot,
death drakes, etc. Fog still fires on every form of taking damage.

Instead, the player is invisible when they are in shadow form (there's
no glow cost for this; the costs of activating/maintaining the ability
are supposed to be a sufficient counterbalance, but they might need to
be adjusted further to make this so).

The transformation power has been halved; on average now it lasts around
70 turns instead of lasting around 100 turns. This may need to be
adjusted downwards further.