Move handling of fiery/illuminating armour to be closer to unrands.
Apart from Chei's hatred of running, no normal armour brand is hated by
a god currently, but plenty of unrands are; it makes sense to be hated
for just wearing the item (as with unrands) instead of merely running
around with them (as with running boots).

Plus, this should allow for more gods to hate particular armour brands
if necessary.