Let shambling mangroves sometimes nest ravens instead of wasps
Intended as part of another overall nerf compared to old treants.

I wouldn't mind seeing a 3rd option for what kind of fauna might
be hiding in their foliage, but I'm not sure what a sensible choice
might be (there seem to be few options that are both thematically
plausible and mechanically reasonable).