Some spriggan band adjustments for Swamp
Druids now have a 1 in 3 chance of coming with a small band of either
plain spriggans or an ocassional rider. Riders themselves now only have
a 1 in 3 chance of their band, and it is smaller. Shambling mangroves
no longer can come with a druid (a hasted mangrove sounds excessive for
a mid-tier Swamp encounter) but sometimes a single rider accompanies
them instead.