Split some prototypes etc off to mon-book.h
All the ones that reference mon_spellbook_type, so that the users
of mon-util.h don't all require a rebuild whenever mon-spll.h is
touched and mon-mst.h regenerated.

For the same reason, change the 'sec' field mon monster_entry back to an
int, and make get_spellbooks a function rather than a method of
monster_info.

I did not split the implementations off into a new .cc file, but that
wouldn't be unreasonable.