Lab revision: "add" threat/loot minivaults, wake up minotaur
While Labyrinths offer a unique threat/layout for portal vaults and
levels as a whole, they have an insubstantial threat in hunger
and tedium of manual travel that potentially takes longer than is
interesting; overall, the non-minotaur parts are homogenous,
uncontrolled, repetitive over many games, and rely on a weak part of
Crawl (not fighting). The flavour and uniqueness of labyrinths,
alongside their age, do make many unwilling to cut them, though.

Thus, I'm using a function to place some vaguely-thematic monsters
of varying numbers and threat in most of the furniture vaults that
already exist for labs, alongside some loot. The monsters cover
rarely-used earth themes, some undead as fallen adventurers, and
other oddities, chosen as vaguely working within the pretext of
a labyrinth. The loot is weighted roughly like the minotaur's stash,
but with gold replacing books, staves, rods, and misc evokables, and
also have no innate quality boosts. Most of the time it's not worth
extending visits to find extra loot without guidance or guarantee
of existence, especially considering the piety clock is most effective
around lab depth, but it potentially gives a little extra ontop of the
normal minotaur fight and rewards.

Also, to allow more surprise and less prep for minotaurs in recognizing
bordered, different-wall-material vaults, and aide them in picking up
their loot to make them distinct threats, nearly-always tag them with
generate_awake. The patrolling behaviour and vault shapes still confine
minotaurs, so it's difficult to outright sneak past.

There's no guarantee that such minivaults even place in a lab at all,
much less that players will find them on the way to the minotaur, so
certainly more should be done to re-invigorate labs. Still, this is
good in lieu of changing vault placement/integration or map size,
especially in giving a little more combat. People may continue to call
for labyrinth removal, but it's important to at least try to improve
them a little, first.