Buff monster merfolk and mermaids somewhat
Merfolk were the most common 'standard' monster in most of Shoals, but
pretty underwhelming - similar in stats to an orc warrior and losing in
melee against a lone elephant, when elephants themselves spawn in packs
from a much earlier floor and aren't even major powerhouses then.

This boosts their damage, hd, and hp, as well as making sure they always
actually spawn with a weapon. Mermaids are now also guaranteed a (lesser)
weapon and get a melee damage boost. Neither is liable to be too much
scarier, but hopefully just a little less ignorable.

(The size of bands involving merfolk were nudged down slightly in
compensation, since the elite merfolk are actually dangerous)