Buff Crypt:$ loot with thematic extra stuff
With the awkward position of late-game optional branch without a rune,
and thus an implied struggle to get people to do the branch when it is
rightfully dangerous but without much of an incentive, Crypt could use
another incentive or two to get people to peek in. Forest, regardless
of its status, at least has sling bullets, lajatangs, quickblades,
and a weird artefact mottled dragon armour; Crypt should mirror this.

Every crypt end now has, in very randomized placement, one of the following:
 * One hall of blades weapon with good pluses and (only) draining brand,
 * One decent but not great weapon with pain or vampiricism, or
 * Two or three randart rings of rN+ or amulets of warding.
Plus, they also have one of the following:
 * Death-themed weapon unrand: Vampire's Tooth, Morg, Doom Knight's, Bloodbane,
   Leech, Zonguldrok, Botono, Finisher, Scythe of Curses, Undeadhunter, or
   the Sceptre of Torment;
 * The necronomicon (1/15 chance) or necro randbook with one necronomicon spell;
 * Evokables stack: one elemental evoker or rod, crystal ball or staff of energy,
   staff of death/power/wizardry/whatever.
Some of these are very valuable, if unreliable, drops, but their placement in a
buffed Crypt should avoid making characters visibly more overpowered when
more characters are killed anyway.

There are also some minor edits in added junk loot for david_glass_crypt and
a reformatting in map defintiion for grunt_crypt_end_deaths_head to adapt to this.