Give water nymphs a passive aura of water which follows them around
This power performs some of the same functions as waterport touch
(forcing melee confrontations to be in range of waterstrike), without
directly throwing the player around. Their new home will have plenty
of water to use anyway, but this serves as a failsafe when lured onto
land (and can also synergize with other nearby creatures).