Guarantee a demonic rune or Pan exit in evilmike_holy_pan.
It's in the heart of the fortress with the main loot pile - you'll need
to get past the seraph to get there.

This has been brought up on several occasions and I think it's worth
trying out for the moment; at the very least it should help get us more
feedback with respect to what to do with the problematic holy monsters.