Place Pan transit gates in minor panlord vaults.
This should give players an incentive to fight through them if they
already have the demonic rune and aren't looking for a way out yet.

There are a few cases where the gate is placed in place of an item; in
those cases I'm placing the item on top of the gate. This may not be
necessary.

Some of the gate placements might be a bit questionable, and these
changes are turning up a lot of vault syntax that could probably be
improved in these vaults; hopefully this should encourage a further look
at the vaults in question.